Added hard coded special IP for global XML file
Added a special XML file with a hard coded “special” IP for a run on
any machine configuration file.  I also modified the XML format to more
semantically define the initiator object and allow binding of
initiators on the same IP.  Made a few small changes to the xmlloader
and writer to support these changes.